+++ sematic/resolvers/BUILD
@@ -40,7 +40,7 @@ sematic_py_lib(
         "//sematic:abstract_future",
         "//sematic:external_resource",
         "//sematic:future_context",
-        "//sematic/resolvers:resource_manager",
+        "//sematic/resolvers/resource_managers:init",
         "//sematic/resolvers:state_machine_resolver",
     ],
 )
@@ -88,14 +88,6 @@ sematic_py_lib(
     name = "resource_requirements",
     srcs = ["resource_requirements.py"],
     deps = [],
-)
-
-sematic_py_lib(
-    name = "resource_manager",
-    srcs = ["resource_manager.py"],
-    deps = [
-        "//sematic:external_resource",
-    ],
 )
 
 sematic_py_lib(
