+++ sematic/resolvers/resource_managers/__init__.py
@@ -1,0 +1,7 @@
+# Sematic
+from sematic.resolvers.resource_managers.in_memory_manager import (  # noqa: F401,E402
+    InMemoryResourceManager,
+)
+from sematic.resolvers.resource_managers.resource_manager import (  # noqa: F401,E402
+    ResourceManager,
+)
